move unit tests to their own file
