+++ graph/lazy_property.py
