+++ setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup, find_packages
+
+setup(name='toplogy', version='2.0', packages=find_packages())
